3000mod/Builds/build2_asyncio/requirements.txt
pyserial>=3.5
pyserial-asyncio>=0.6
PyYAML>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
